+++ addOns/ascanrules/CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 - The Format String Error scan rule now includes example alert functionality for documentation generation purposes (Issue 6119).
 
+### Changed
+- Move time based attacks to their own scan rules (Issue 7341).
+
 ## [55] - 2023-06-06
 ### Changed
 - The Parameter Tamper Scan rule now includes example alert functionality for documentation generation purposes (Issue 6119)
